Homo sapiens (human)
This gene encodes a divergent member of the CD1 family of transmembrane glycoproteins, which are structurally related to the major histocompatibility complex (MHC) proteins and form heterodimers with beta-2-microglobulin. The CD1 proteins mediate the presentation of primarily lipid and glycolipid antigens of self or microbial origin to T cells. The human genome contains five CD1 family genes organized in a cluster on chromosome 1. The CD1 family members are thought to differ in their cellular localization and specificity for particular lipid ligands. The protein encoded by this gene localizes to late endosomes and lysosomes via a tyrosine-based motif in the cytoplasmic tail. Two transcript variants encoding different isoforms have been found for this gene. [provided by RefSeq, Jan 2016]